fix #23: separate modules for endpoints and framework
