+++ src/framework/endpoint.rs
@@ -1,5 +1,5 @@
-use super::Environment;
-use crate::response::ApiResult;
+use crate::framework::response::ApiResult;
+use crate::framework::Environment;
 use serde::Serialize;
 use url::Url;
